// web/src/generated/api/models/BalanceParticipantResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cuentas Claras API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * One stable-order participant balance row.
 * @export
 * @interface BalanceParticipantResponse
 */
export interface BalanceParticipantResponse {
    /**
     * 
     * @type {boolean}
     * @memberof BalanceParticipantResponse
     */
    archived: boolean;
    /**
     * 
     * @type {number}
     * @memberof BalanceParticipantResponse
     */
    balanceCents: number;
    /**
     * 
     * @type {string}
     * @memberof BalanceParticipantResponse
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof BalanceParticipantResponse
     */
    owedCents: number;
    /**
     * 
     * @type {number}
     * @memberof BalanceParticipantResponse
     */
    paidCents: number;
    /**
     * 
     * @type {string}
     * @memberof BalanceParticipantResponse
     */
    participantId: string;
}

/**
 * Check if a given object implements the BalanceParticipantResponse interface.
 */
export function instanceOfBalanceParticipantResponse(value: object): value is BalanceParticipantResponse {
    if (!('archived' in value) || value['archived'] === undefined) return false;
    if (!('balanceCents' in value) || value['balanceCents'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('owedCents' in value) || value['owedCents'] === undefined) return false;
    if (!('paidCents' in value) || value['paidCents'] === undefined) return false;
    if (!('participantId' in value) || value['participantId'] === undefined) return false;
    return true;
}

export function BalanceParticipantResponseFromJSON(json: any): BalanceParticipantResponse {
    return BalanceParticipantResponseFromJSONTyped(json, false);
}

export function BalanceParticipantResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): BalanceParticipantResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'archived': json['archived'],
        'balanceCents': json['balance_cents'],
        'name': json['name'],
        'owedCents': json['owed_cents'],
        'paidCents': json['paid_cents'],
        'participantId': json['participant_id'],
    };
}

export function BalanceParticipantResponseToJSON(json: any): BalanceParticipantResponse {
    return BalanceParticipantResponseToJSONTyped(json, false);
}

export function BalanceParticipantResponseToJSONTyped(value?: BalanceParticipantResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'archived': value['archived'],
        'balance_cents': value['balanceCents'],
        'name': value['name'],
        'owed_cents': value['owedCents'],
        'paid_cents': value['paidCents'],
        'participant_id': value['participantId'],
    };
}